requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Flask-CORS==4.0.0
reportlab==4.4.3
Werkzeug==2.3.7

